+++ src/main/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/AbstractCompoundAssetFilter.java
@@ -11,12 +11,12 @@ import java.util.Set;
 @JsonTypeInfo(use = JsonTypeInfo.Id.DEDUCTION)
 public abstract class AbstractCompoundAssetFilter implements AssetFilter {
     @Override
-    public final Set<String> apply(Map<String, Asset> allAssets, Set<String> currentSelection) {
+    public final Map<String, Asset> apply(Map<String, Asset> input) {
         var constituentFilters = getConstituentFilters();
         Preconditions.checkState(constituentFilters.size() > 0, "Compound filter requires >= 1 constituent filter");
 
         var filterResultsOpt = getConstituentFilters().stream()
-                .map(filter -> filter.apply(allAssets, currentSelection))
+                .map(filter -> filter.apply(input))
                 .reduce(this::combineFilterMatches);
 
         // We're okay to ignore the Optional.isPresent() check because we assert that the constituent filters
@@ -28,5 +28,5 @@ public abstract class AbstractCompoundAssetFilter implements AssetFilter {
     protected abstract List<AssetFilter> getConstituentFilters();
 
     // Reduce function used for deciding
-    protected abstract Set<String> combineFilterMatches(Set<String> filterResultA, Set<String> filterResultB);
+    protected abstract Map<String, Asset> combineFilterMatches(Map<String, Asset> filterResultA, Map<String, Asset> filterResultB);
 }