+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/DisjunctionAssetFilterTest.java
@@ -57,9 +57,13 @@ public class DisjunctionAssetFilterTest {
                 needle2Filter
         ));
 
-        var result = disjunctionFilter.apply(haystack, haystack.keySet());
+        var result = disjunctionFilter.apply(haystack);
 
-        var expected = Set.of(matchingAssetId1, matchingAssetId2, matchingAssetId3);
+        var expected = Map.of(
+                matchingAssetId1, haystack.get(matchingAssetId1),
+                matchingAssetId2, haystack.get(matchingAssetId2),
+                matchingAssetId3, haystack.get(matchingAssetId3)
+        );
         Assert.assertEquals(expected, result);
     }
 }