+++ src/test/java/com/strangegrotto/wealthdraft/assetallocation/datamodel/filters/TagAssetFilterTest.java
@@ -37,8 +37,11 @@ public class TagAssetFilterTest {
         );
 
         var filter = ImmTagAssetFilter.of(customTags, selectedTagName, selectedTagValue);
-        var result = filter.apply(assets, assets.keySet());
-        var expected = Set.of(matchingAssetId1, matchingAssetId2);
+        var result = filter.apply(assets);
+        var expected = Map.of(
+                matchingAssetId1, assets.get(matchingAssetId1),
+                matchingAssetId2, assets.get(matchingAssetId2)
+        );
         Assert.assertEquals(expected, result);
     }